requirements.txt
flask==3.0.2
tensorflow-cpu==2.15.0
gunicorn==21.2.0
pillow==10.2.0
numpy==1.24.3
gdown==4.7.1
py7zr==0.20.5
Flask-CORS==3.0.10
